pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralph-orchestrator"
version = "0.1.0"
description = "Run-anywhere autonomous workflow orchestrator (Ralph pattern + verified gates + autopilot)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
  "jsonschema>=4.23.0",
]

[project.scripts]
ralph = "ralph_orchestrator.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["ralph_orchestrator*"]
